move qt/vispy event conversion under the qt_viewer
